Revert "Revert "Make menubar its own widget and add overflow (#63954)""
This reverts commit fa87e67ebfd6b8014ba697ddbc6c2d6c0e2cb970.
